requirements.txt
flask
flask-sqlalchemy
psycopg2-binary==2.9.9
qdrant-client
google-generativeai>=0.4.1,<0.5.0
python-dotenv
flask-cors
langchain
langchain-google-genai
langchain-community
fastapi
langchain_groq
sqlalchemy
uvicorn
reportlab